+++ src/framework/execution/logger.ts
@@ -11,7 +11,11 @@ import {
 } from './types';
 
 import { SynchronizationJobContext } from '../synchronization';
-import { UNEXPECTED_ERROR_CODE } from '../../errors';
+import {
+  IntegrationError,
+  UNEXPECTED_ERROR_CODE,
+  UNEXPECTED_ERROR_REASON,
+} from '../../errors';
 
 // eslint-disable-next-line
 const bunyanFormat = require('bunyan-format');
@@ -211,6 +215,16 @@ function instrumentEventLogging(
 
       publishEvent(name, description);
     },
+    validationFailure: (err: Error) => {
+      const name = 'validation-failure';
+      const { errorId, description } = createErrorEventDescription(
+        err,
+        `Error occurred while validating integration configuration.`,
+      );
+
+      logger.error({ errorId, err }, description);
+      publishEvent(name, description);
+    },
     child: (options: object = {}, simple?: boolean) => {
       const c = child.apply(logger, [options, simple]);
       return instrumentEventLogging(
@@ -222,12 +236,27 @@ function instrumentEventLogging(
   });
 }
 
-function createErrorEventDescription(err: Error, message: string) {
+export function createErrorEventDescription(
+  err: Error | IntegrationError,
+  message: string,
+) {
   const errorId = uuid();
-  const errorCode = (err as any).code ?? UNEXPECTED_ERROR_CODE;
+
+  let errorCode: string;
+  let errorReason: string;
+
+  if (err instanceof IntegrationError) {
+    errorCode = err.code;
+    errorReason = err.message;
+  } else {
+    errorCode = UNEXPECTED_ERROR_CODE;
+    errorReason = UNEXPECTED_ERROR_REASON;
+  }
+
+  const errorDetails = `errorCode=${errorCode}, errorId=${errorId}, reason=${errorReason}`;
 
   return {
     errorId,
-    description: `${message} (errorCode=${errorCode}, errorId=${errorId})`,
+    description: `${message} (${errorDetails})`,
   };
 }